+++ R/reporter-check.R
@@ -10,84 +10,77 @@ CheckReporter <- R6::R6Class("CheckReporter",
   public = list(
     problems = NULL,
     skips = NULL,
-
+    warnings = NULL,
     n_ok = 0L,
-    n_skip = 0L,
-    n_fail = 0L,
-    n_warn = 0L,
 
     initialize = function(...) {
       self$capabilities$parallel_support <- TRUE
       self$problems <- Stack$new()
+      self$warnings <- Stack$new()
       self$skips <- Stack$new()
 
       super$initialize(...)
     },
 
     add_result = function(context, test, result) {
-      if (expectation_success(result)) {
-        self$n_ok <- self$n_ok + 1L
-        return()
-      }
-
-      if (!expectation_skip(result)) {
+      if (expectation_broken(result)) {
         self$problems$push(result)
-      }
-
-      if (expectation_skip(result)) {
-        self$n_skip <- self$n_skip + 1L
-        self$skips$push(result$message)
       } else if (expectation_warning(result)) {
-        self$n_warn <- self$n_warn + 1L
+        self$warnings$push(result)
+      } else if (expectation_skip(result)) {
+        self$skips$push(result$message)
       } else {
-        self$n_fail <- self$n_fail + 1L
+        self$n_ok <- self$n_ok + 1L
       }
-
-      type <- expectation_type(result)
-      msg <- issue_header(result)
-      header <- cli::rule(msg, width = max(nchar(msg) + 6, 80))
-
-      self$local_user_output()
-      self$cat_line(header)
-      if (expectation_warning(result)) {
-        # A lot of CRAN packages have warnings so showing full traceback
-        # makes it hard to see failures
-        self$cat_line(result$message)
-      } else {
-        self$cat_line(format(result, simplify = "none"))
-      }
-      self$cat_line()
     },
 
     end_reporter = function() {
-      if (self$n_skip > 0) {
-        self$rule("Skipped tests ", line = 1)
+      if (self$skips$size() > 0) {
+        self$rule("Skipped tests", line = 2)
         self$cat_line(skip_bullets(self$skips$as_list()))
         self$cat_line()
       }
 
-      problems <- self$problems$as_list()
-      if (length(problems) == 0) {
-        return()
+      # Don't show warnings in revdep checks in order to focus on failures
+      if (self$warnings$size() > 0 && !is_revdep_check()) {
+        warnings <- self$warnings$as_list()
+
+        self$rule("Warnings", line = 2)
+        self$cat_line(map_chr(warnings, issue_summary, rule = TRUE))
+        self$cat_line()
       }
 
-      saveRDS(problems, "testthat-problems.rds")
-      self$rule("testthat results ", line = 2)
-      self$cat_line(vapply(problems, issue_header, character(1)))
-      self$cat_line()
-      self$cat_line(summary_line(self$n_fail, self$n_warn, self$n_skip, self$n_ok))
+      if (self$problems$size() > 0) {
+        problems <- self$problems$as_list()
+        saveRDS(problems, "testthat-problems.rds")
+
+        self$rule("Failed tests", line = 2)
+        self$cat_line(map_chr(problems, issue_summary, rule = TRUE, simplify = "none"))
+        self$cat_line()
+      }
+
+      self$cat_line(summary_line(
+        n_fail = self$problems$size(),
+        n_warn = self$warnings$size(),
+        n_skip = self$skips$size(),
+        n_pass = self$n_ok
+      ))
     }
   )
 )
 
-summary_line <- function(n_fail, n_warn, n_skip, n_ok) {
+summary_line <- function(n_fail, n_warn, n_skip, n_pass) {
   # Ordered from most important to least important
   paste0(
     "[ ",
     colourise("FAIL", "failure"), " ", n_fail, " | ",
     colourise("WARN", "warn"),    " ", n_warn, " | ",
     colourise("SKIP", "skip"),    " ", n_skip, " | ",
-    colourise("PASS", "success"), " ", n_ok,
+    colourise("PASS", "success"), " ", n_pass,
     " ]"
   )
+}
+
+is_revdep_check <- function() {
+  !identical(Sys.getenv("DEV_PACKAGE_NAME"), "")
 }